Remove unnecessary type casts.

+++ src/main/java/org/apache/commons/codec/digest/MurmurHash2.java
@@ -166,7 +166,7 @@ public final class MurmurHash2 {
 		case 2:
 			h ^= (long) (data[(length & ~7) + 1] & 0xff) << 8;
 		case 1:
-			h ^= (long) (data[length & ~7] & 0xff);
+			h ^= data[length & ~7] & 0xff;
 			h *= m;
 		}
